refactor(goals): Decimal money em DTOs + goal_service math (A6g.3b slice 2 · ADR-090)
11 campos money migrados em aporte/dolar/if_goal (MoneyBRL/MoneyUSD),
matemática do goal_service reescrita em Decimal (ln/exp p/ taxa
mensal), persistência via model_dump(mode="json") mantém JSON number
on disk. OpenAPI snapshot atualizado (Input/Output split por causa
do PlainSerializer — wire Output permanece number puro, frontend TS
intacto). 1 assert de teste convertido para float(meta) na aritmética
mista com r_m float.

+++ backend/app/application/goal/get_active_if_goal.py
@@ -2,6 +2,7 @@
 
 from __future__ import annotations
 
+from decimal import Decimal
 from typing import Optional
 
 from backend.app.application.base.errors import NotFoundError
@@ -14,7 +15,7 @@ async def get_active_if_goal(
     workspace_id: str,
     *,
     repo: GoalRepositoryProtocol,
-    patrimonio_atual_brl: Optional[float] = None,
+    patrimonio_atual_brl: Optional[Decimal] = None,
     created_by_name: Optional[str] = None,
 ) -> IFGoalResponse:
     """Retorna meta IF com ``derived`` recalculado usando patrimônio atual.
